SlotAnnotator class converted to abstract class
