+++ doc/source/api/agent.rst
@@ -3,3 +3,10 @@ Agent
 
 .. automodule:: moviebot.agent.agent
    :members:
+
+.. toctree::
+    :caption: Components:
+
+    nlu
+    dialogue_manager
+    nlg
