+++ doc/source/api/nlu.annotators.rst
@@ -1,0 +1,11 @@
+Slot Annotators
+===============
+
+.. automodule:: moviebot.nlu.slot_annotator
+   :members:
+
+Rule-based annotator
+--------------------
+
+.. automodule:: moviebot.nlu.rule_based_annotator
+    :members:
