+++ doc/source/api/nlu.intent_checker.rst
@@ -1,0 +1,5 @@
+Intent Checker
+===============
+
+.. automodule:: moviebot.nlu.user_intents_checker
+   :members:
