+++ doc/source/index.rst
@@ -4,10 +4,10 @@
    contain the root `toctree` directive.
 
 IAI MovieBot |release| documentation
-===================
+=====================================
 
 IAI MovieBot is an open-source, conversational movie recommender system which models users' preferences dynamically and supports user initiatives and multi-turn recommendations.
-IAI MovieBot equips with a scaleable structure for future amendments. It facilities the standardized components.
+IAI MovieBot equips with a scalable structure for future amendments. It facilities the standardized components.
 In specific, intent recognition in NLU identifies users' intent based on preferences and recognizes entities (movies and attributes) for users' utterances, NLG generates natural language responses based on templates, and dialogue policy in the dialogue manager adapts to the users' requirements in the conversation.
 
 .. toctree::
