+++ moviebot/intents/user_intents.py
@@ -11,22 +11,22 @@ class UserIntents(Enum):
     """This class contains the list of possible intents for user as constants.
     This approach will help the NLU to identify the intents from a pre-defined set."""
 
-    REVEAL = "reveal"
+    REVEAL = 'reveal'
 
-    INQUIRE = "inquire"
-    REMOVE_PREFERENCE = "remove_preference"
+    INQUIRE = 'inquire'
+    REMOVE_PREFERENCE = 'remove_preference'
 
-    REJECT = "reject"
-    ACCEPT = "accept"
+    REJECT = 'reject'
+    ACCEPT = 'accept'
 
-    CONTINUE_RECOMMENDATION = "continue_recommendation"
-    RESTART = "restart"
+    CONTINUE_RECOMMENDATION = 'continue_recommendation'
+    RESTART = 'restart'
 
-    UNK = "UNK"
-    ACKNOWLEDGE = "acknowledge"
-    DENY = "deny"
+    UNK = 'UNK'
+    ACKNOWLEDGE = 'acknowledge'
+    DENY = 'deny'
     HI = 'hi'
-    BYE = "bye"
+    BYE = 'bye'
 
     def __str__(self):
         return self.value
